consolidate test engine setup logic into helpers
